+++ GRDB/Core/DatabaseError.swift
@@ -32,7 +32,7 @@ extension DatabaseError: CustomStringConvertible {
         if let sql = sql {
             description += " with statement `\(sql)`"
         }
-        if let arguments = arguments {
+        if let arguments = arguments where !arguments.isEmpty {
             description += " arguments \(arguments)"
         }
         if let message = message {